requirements.txt
eve
dnspython
flask-cors
python-dotenv